+++ tests/test_core.py
@@ -119,6 +119,35 @@ def test_from_dict_source__list_nested():
     )
 
 
+def test_from_dict_source__tuple():
+    from typing import Tuple
+
+    @dataclass
+    class A:
+        a: str
+
+    @dataclass
+    class B:
+        b: str
+
+    @dataclass_json
+    @dataclass
+    class C:
+        c: Tuple[A, B]
+
+    assert core._from_dict_source(C) == textwrap.dedent(
+        """\
+        def from_dict(cls, o, *, infer_missing):
+            args = {}
+            value = o.get('c')
+            if value is not None:
+                value = (__0:=(value),(A._fastclasses_json_from_dict(__0[0]),B._fastclasses_json_from_dict(__0[1]),))[1]
+            args['c'] = value
+            return cls(**args)
+        """
+    )
+
+
 def test_from_dict_source__enum():
     from enum import Enum