+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for all kinds of keys types that `json.dumps` supports, for `Dict`
   fields. And UUIDs
-- Support for Mapping and Sequence
+- Support for Mapping, Sequence and Tuple fields
 
 ## [0.3.2] - 2021-12-07
 ### Fixed